Set Traceur notifications as blockable
This change explicitly sets Traceur's notification channels as
blockable. There previously appeared to be some by-user or by-device
variations in whether or not Traceur notifications could be modified.

For example, the reporter of the bug below could not modify their
device's system tracing notifications, but wiping+flashing my device
to the same build didn't allow me to repro.

Test: Use device that previously had unmodifiable Traceur notifications
      to verify that the notifications can be edited.
Bug: 241003757
Change-Id: I0849cdd279d4043fd4095e3f69e89a4b974581b9