Add CPU profiling to Traceur
This change adds the ability to sample callstacks in Traceur. A
dedicated toggle for capturing only callstacks has been added to the
main app UI, and a "cpu" category has been added so that callstack
sampling can be enabled for the existing trace function.

This change also genericizes "traces" to "files" or "recordings" where
appropriate.

Test: Check the following:
      - "Record CPU profile" collects a trace with callstacks that are
        viewable in the Perfetto UI (with process names).
      - Saved stack samples are saved with the "stack-samples" prefix.
      - "Clear saved files" also clears stack samples.
      - Tapping the ongoing-samples notification stops recording.
      - "Record trace" with the "cpu" category checked collects a trace
        with callstack samples.
      - "Record trace" and "Record CPU profile" cannot both be enabled
        (either through the main app UI or the QS tile).
Bug: 265048436
Change-Id: Id05bc8db53bbc548d15bed21adeaa6f5285292f7