Update Traceur tests with scrolls + profiling test
This change updates TraceurAppTests to scroll to find elements when
necessary, as there are now enough options/toggles that it cannot be
assumed that all elements are always visible.

This change also adds a CPU profiling test, which is identical in
function to the existing trace test (start and stop recording; attempt
to share it).

A few other changes are also included:
- Some functions for the recording test have been refactored to use in
  the profiling test.
- The UIAutomator library now uses the non-deprecated AndroidX version.
- Occurrences of the "find element in-app by text; click" steps have
  been replaced with findObjectOnMainScreenByText(), which handles
  scrolling as well.

Test: atest TraceurUiTests
Bug: 265048436
Change-Id: Ie9fd5af5b9a164d44dd2c3ffe73091566a3ba3c1