Update Traceur to check admin user status
This change updates Traceur to check for admin user privileges wherever
a developer options check occurs. This is intended to address the case
in which developer options (a global setting not differentiated on
current user privileges) being enabled would allow guest users to open
Traceur through a 3P app and view its trace files. This would previously
be possible even when ADB debugging was disabled by the admin user.

Traceur now listens for user changes so that its document root
(containing traces) is enabled/disabled based on the new user's admin
status.

In S, checking for admin-user status requires the MANAGE_USERS
permission (instead of QUERY_USERS, which exists in newer branches). In
addition, Traceur is a priv-app in S. Because of this, the MANAGE_USERS
permission is explicitly requested by Traceur, and a separate change to
the privapp permissions allowlist is required.

Test: Using ABTD, apply this change to sc-v2-dev and check that:
      - There are no merge conflicts
      - CtsIntentSignatureTestCases passes (b/270791503)
      - TraceurUiTests passes
      Build+flash a local device on sc-v2-dev and check that:
      - Traceur cannot be opened through 'am start' on a guest account
      - Opening Files on a guest account no longer shows a System Traces
        folder (even if Traceur's onCreate is somehow called)
      - System tracing no longer appears in settings for guests
Bug: 262243665
Bug: 262244249
Ignore-AOSP-First: Internal-first security fix
Change-Id: I8bdd42d4e5346e8a71f7bbbb6b53f729f66d4d57
Merged-In: I1c0c8c9588554378ae39a1a69a35ff44052b93e0