Manually revert Traceur admin status change
This change manually reverts the "Update Traceur to check admin user
status" change, as the original revert wasn't automerged to this branch
(or its children) correctly.

The files changed in this CL were copied over from a tm-dev checkout of
the packages/apps/Traceur project, as tm-dev is in the correct state.
This should be safe, as no other changes had been made to Traceur in
tm-dev after the revert.

This change should /not/ automerge into tm-qpr-dev (or further), because
tm-qpr-dev is currently in the correct state (so this change would find
a conflict). The Merged-In tag being used below has been taken from the
most recent automerger change's Change-Id on tm-qpr-dev, as this
Change-Id is guaranteed to be unique.

Test: presubmit
Bug: 271493457
Merged-In: I2b510f1e6881882879ccd64ad8f8b5c1affac3e9
Ignore-AOSP-First: Internal-only branch fix.
Change-Id: I0f56a93ef66f4289f98828200e21ae2204bc895c