Manually revert Traceur admin status change
This change manually reverts the "Update Traceur to check admin user
status" change, as the original revert wasn't automerged to this branch
(or its children) correctly.

The files changed in this CL were copied over from a tm-dev checkout of
the packages/apps/Traceur project, as tm-dev is in the correct state.
This should be safe, as no other changes had been made to Traceur in
tm-dev after the revert.

This change should /not/ automerge into udc-dev (or further), because
udc-dev is currently in the correct state (so this change would find a
conflict). The Merged-In tag being used below has been taken from the
most recent automerger change's Change-Id on udc-dev, as this Change-Id
is guaranteed to be unique.

Test: presubmit
Bug: 271493457
Merged-In: Iac8cc2b8ff12feb640668623b3137c85ee73bd86
Ignore-AOSP-First: Internal-only branch fix.
Change-Id: I22bc6079d2936438989a6fe65340cccd14d6886e