Manually revert Traceur admin status change
This change manually reverts the "Update Traceur to check admin user
status" change, as the original revert wasn't automerged to this branch
(or its children) correctly.

The files changed in this CL were copied over from a udc-dev checkout of
the packages/apps/Traceur project, as udc-dev is in the correct state.
This should be safe, as no other changes had been made to Traceur in
udc-dev after the revert.

This change should /not/ automerge into master, because master is
currently in the correct state (so this change would find a conflict).
The Merged-In tag being used below has been taken from the most recent
automerger change's Change-Id on master, as this Change-Id is guaranteed
to be unique.

Test: presubmit
Bug: 271493457
Merged-In: I9d10bec74d0bb8c68a323797feba5d5d7191fcff
Ignore-AOSP-First: Internal-only branch fix.
Change-Id: Ia92f06d9d481e47372afaf7e315f7bd29b49c0e4