Add QS tile for stack samples
This change adds the option to enable a QS tile for stack sampling / CPU
profiling. The toggle for enabling this tile has intentionally been
positioned _above_ the existing tracing QS tile toggle in order to avoid
confusing users who are used to seeing that at the bottom of the page.

There is a bug where one tile being tapped to stop a recording doesn't
appear to make the other tile available. This is just a UI bug though,
as the other tile can still be interacted with, and refreshing the quick
settings bar makes the other tile appear available again.

Test: Check that Traceur's QS tile(s) still work as expected:
      - they reflect the state of tracing
      - they can be used to stop/start recordings
      - they collect the correct type of recordings
Bug: 326869302
Change-Id: I75b0e6bb4d3962d0ce625080a05303380476f662