Remove TraceEngine interface definition
The TraceEngine interface was an artifact from when there were two
tracing backends available (atrace and Perfetto). Only Perfetto is
supported now, so the interface is no longer necessary. This eliminates
the need for a function definition to be added to TraceEngine every time
a new PerfettoUtils function is called from TraceUtils.

Test: Build, flash, check that tracing works as expected.
Bug: 306420759
Change-Id: I37173af639a4b0a3d22c52d71ba545a097f1f134