Remove InternalReceiver
InternalReceiver was originally added to give users the ability to start
new traces directly from the notification that appeared after a previous
trace session was stopped and stolen by a bug report.

As of aosp/2323319, trace sessions are no longer stolen and are cloned
instead, meaning that InternalReceiver can be removed.

Test: Check in both AOSP and internal-main builds that there are no
      differences before/after this change when terminating a trace
      by capturing a bug report.
Bug: 352779024
Change-Id: Ia21fa8a5a9ab1a23985745ec8b02cc61cf758bc2