Update share sheet to use Intent.createChooser()
This change wraps the trace-sharing intent (accessed through the
notification shown after a trace is saved) in Intent.createChooser().
This is the method recommended in public Android docs, and is intended
to fix some inconsistencies with other share sheets in Android.

This has the effect of removing the "Just once" and "Always" options for
share destinations.

Test: - Check that the share sheet is still functional on devices with
        and without a work profile.
      - Check that the share sheet is now in line with the share sheet
        for other data types (photos, links, etc.).
Bug: 308617304
Change-Id: I64ef747901219d1b772e69a69ae86adf1d8084ea