Remove SystemUser requirement from BindableTraceService.
This is still being guarded by a permission. However, there was a crash
in Headless System User Mode that was caused by this restriction. It
isn't necessary for protection so it is being removed.

Bug: 364824477
Test: Compilation worked after removing this line from the android
manifest. As part of a greater effort (with other necessary CLs), I verified that this change has the added effect of enabling com.android.systemui to connect to it from a different, non-system user's profile. The test was to use the record issue tile as intended, with good interactions between apps, processes, and users.
Flag: EXEMPT bug fix
Ignore-AOSP-First: This class isn't in AOSP yet.

Change-Id: I05894413195be2e674b743337b5cecba35356655