Update TraceController to always reply when sharing
This change updates TraceController to respond with an empty message
when there are no trace files to share. This is intended to improve
debuggability by making it easier to identify where in the
Perfetto->Traceur->client flow an error has occurred.

Bug: 364824477
Test: Verified everything still works normally.
Flag: EXEMPT small refactor
Ignore-AOSP-First: This feature is not yet in AOSP.
Change-Id: I29aadd8283d149918d4868be43f98b7fb01ee150